requirements.txt
GitPython
python-gitlab
PyGithub
matplotlib
notebook
python-sonarqube-api
docker
Unidecode